+++ lib/controller.dart
@@ -28,6 +28,8 @@ export 'package:mvc_pattern/mvc_pattern.dart' show AppMVC, AppConMVC;
 export 'package:mvc_application/src/controller/app.dart'
     show AppController, ErrorHandler, ControllerMVC;
 
+export 'package:mvc_application/src/controller/alarm_manager.dart' show AlarmManager;
+
 export 'package:mvc_application/src/controller/deviceinfo.dart' show DeviceInfo;
 
 export 'package:mvc_application/src/controller/assets/assets.dart';